极大规模地重新组织了代码 修改了主窗口hwnd获取的相关代码，使其更加简洁 及其他部分bug/逻辑含糊之处
