+++ Gui/Widgets.py
@@ -1,66 +1,418 @@
 # -*- coding: UTF-8 -*-
 import re
+import tkinter as tk
 from base64 import b64decode, b64encode
-from json import load as json_load, dump as json_dump, JSONDecodeError
-from os.path import exists, join as path_join
+from copy import copy
+from json import load as json_load, JSONDecodeError, dump as json_dump
+from math import ceil
+from os.path import join as path_join, exists
 from pickle import loads as pickle_loads, dumps as pickle_dumps
 from queue import Queue
+from random import randint
 from sys import stderr
-from threading import Thread, Lock
-from time import perf_counter, sleep, time, strftime, localtime
-from tkinter import font, filedialog
-from typing import Dict
+from threading import Lock, Thread
+from time import strftime, localtime, time, sleep, perf_counter
+from tkinter import Misc, filedialog
+from tkinter.font import Font
+from typing import Any, List, Dict
 
+import ttkbootstrap as ttk
+from PIL import ImageDraw, ImageTk, Image
 from comtypes import CoInitialize, CoUninitialize
 from ping3 import ping
-from pyperclip import copy as copy_clipboard
+from pyperclip import copy as copy_clipboard, copy
+from ttkbootstrap import Style
+from ttkbootstrap.constants import *
 from ttkbootstrap.scrolled import ScrolledFrame
-from win32con import (MB_ICONQUESTION,
-                      MB_ICONWARNING,
-                      MB_YESNOCANCEL,
-                      MB_ICONERROR,
-                      MB_YESNO,
-                      IDCANCEL,
-                      MB_OK,
-                      IDYES,
-                      IDNO)
-from win32gui import (EnumChildWindows,
-                      GetWindowText,
-                      SetWindowText,
-                      FindWindowEx,
-                      GetClassName,
-                      FindWindow,
-                      MessageBox,
-                      GetParent)
+from win32con import MB_OK, MB_ICONERROR, MB_YESNOCANCEL, MB_ICONQUESTION, IDYES, IDNO, IDCANCEL, MB_ICONWARNING, \
+    MB_YESNO
+from win32gui import GetWindowText, SetWindowText, FindWindow, FindWindowEx, GetParent, EnumChildWindows, MessageBox, \
+    GetClassName
 
-from info_gui import InfoWindow
-from scanner import ServerScanner
-from taskbar_lib import *
-from vars import config_dir
-from widgets import *
+from Gui.InfoGui import InfoWindow
+from Libs.ColorLib import Color
+from Libs.TaskbarLib import TaskbarApi, TBPFLAG
+from Libs.Vars import scale_rater, config_dir, UserAddressOperator, server_addresses
+from Network.Scanner import ServerInfo, ServerScanner
 
+ERROR = "error"
 DEBUG = "debug"
 scanbar: Any = None
 
 
+class MOTD(ttk.Text):
+    def __init__(self, master: tk.Misc):
+        super(MOTD, self).__init__(master, state=DISABLED, height=1, width=70, relief=FLAT)
+
+    def load_motd(self, data: ServerInfo):
+        self.configure(state=NORMAL)
+        try:
+            self.delete("1.0", END)
+        except IndexError:
+            pass
+        for extra in data.description_json:
+            try:
+                tag = hex(randint(0, 114514))
+                tag_font = Font(family="Unifont", size=12)
+                if extra.get("color"):
+                    if "#" not in extra["color"]:
+                        color = vars.color_map_hex[extra["color"]]
+                    else:
+                        color = extra["color"]
+                    self.tag_configure(tag, foreground=color)
+
+                if extra.get("underline") or extra.get("underlined"):
+                    self.tag_configure(tag, underline=True)
+                if extra.get("bold"):
+                    tag_font.config(family="宋体", weight="bold")
+                elif extra.get("italic"):
+                    tag_font.config(slant="italic")
+                elif extra.get("strikethrough"):
+                    tag_font.config(overstrike=True)
+
+                self.tag_configure(tag, font=tag_font, justify=LEFT)
+                self.insert(END, extra["text"], tag)
+            except TimeoutError as e:
+                print("MOTD Data Extra Error:", extra, e)
+        self.configure(state=DISABLED)
+
+
+class EntryScale(ttk.Frame):
+    def __init__(self, master: Misc, _min: Any, _max: Any, value: Any, text: str, fmt: Any):
+        super(EntryScale, self).__init__(master)
+        self.min = _min
+        self.max = _max
+        self.fmt = fmt
+        self.text = ttk.Label(self, text=text)
+        self.entry = ttk.Entry(self, width=8)
+        self.scale = ttk.Scale(self, from_=_min, to=_max, command=self.scale_set_value)
+
+        self.entry.bind("<KeyRelease>", self.entry_set_value)
+        self.entry.insert(END, str(value))
+        self.entry_set_value()
+
+        self.text.pack(side=LEFT)
+        self.scale.pack(side=LEFT, fill=X, expand=True)
+        self.entry.pack(side=LEFT)
+
+    def scale_set_value(self, _=None):
+        value = self.fmt(self.scale.cget("value"))
+        if isinstance(value, float):
+            value = round(value, 2)
+        self.scale.configure(value=value)
+        self.entry.delete(0, END)
+        self.entry.insert(0, str(value))
+
+    def entry_set_value(self, _=None):
+        value = self.entry.get()
+        try:
+            value = self.fmt(value)
+            if value <= self.max:
+                if value >= self.min:
+                    self.scale.configure(value=value)
+                else:
+                    self.entry.delete(0, END)
+                    self.entry.insert(0, str(self.min))
+                    self.entry_set_value()
+            else:
+                self.entry.delete(0, END)
+                self.entry.insert(0, str(self.max))
+                self.entry_set_value()
+        except ValueError:
+            pass
+
+    def get_value(self):
+        return self.fmt(self.scale.cget("value"))
+
+    def set_value(self, value):
+        self.scale.configure(value=value)
+        self.entry.delete(0, END)
+        self.entry.insert(0, str(value))
+
+
+class EntryScaleInt(EntryScale):
+    def __init__(self, master: Misc, _min: int, _max: int, value: int, text: str):
+        super(EntryScaleInt, self).__init__(master, _min, _max, value, text, int)
+
+
+class EntryScaleFloat(EntryScale):
+    def __init__(self, master: Misc, _min: float, _max: float, value: float, text: str):
+        super(EntryScaleFloat, self).__init__(master, _min, _max, value, text, float)
+
+
+class TextEntry(ttk.Frame):
+    def __init__(self, master: Misc, tip: str, value: str = ""):
+        super(TextEntry, self).__init__(master)
+
+        self.text = ttk.Label(self, text=tip)
+        self.entry = ttk.Entry(self)
+
+        self.entry.insert(0, value)
+
+        self.text.pack(side=LEFT)
+        self.entry.pack(side=LEFT, fill=X, expand=True)
+
+    def get(self):
+        return self.entry.get()
+
+    def delete(self, first, last) -> None:
+        self.entry.delete(first, last)
+
+    def set(self, text: str):
+        self.entry.delete(0, END)
+        self.entry.insert(0, text)
+
+
+class TextCombobox(ttk.Frame):
+    def __init__(self, master: Misc, tip: str, value: List[str] = None):
+        super(TextCombobox, self).__init__(master)
+
+        self.text = ttk.Label(self, text=tip)
+        self.combobox = ttk.Combobox(self, values=value)
+
+        if len(value) > 0:
+            self.combobox.current(0)
+
+        self.text.pack(side=LEFT)
+        self.combobox.pack(side=LEFT, fill=X, expand=True)
+
+    def get(self) -> str:
+        return self.combobox.get()
+
+    def set(self, text: str):
+        self.combobox.delete(0, END)
+        self.combobox.insert(0, text)
+
+
+class Tabs(ttk.Notebook):
+    pass
+
+
+class RangeScale(ttk.Canvas):
+    def __init__(self, master: Misc):
+        """
+        范围选择条
+        使用set设置范围, 使用value获取当前值
+        绑定<<RangeChanged>>事件侦测范围变化
+        """
+        super(RangeScale, self).__init__(master, height=int(scale_rater() * 15))
+        self.bind("<Button-1>", self.mouse_down)
+        self.bind("<ButtonRelease-1>", self.mouse_up)
+        self.bind("<Configure>", self.redraw)
+        self.bind("<Motion>", self.mouse_move)
+        self.bind("<Leave>", self.mouse_move)
+        self.bind("<<ThemeChanged>>", self.color_change)
+
+        self.min_percentage: float = 0.25
+        self.max_percentage: float = 0.75
+
+        self.image = None
+        self.image_tk = None
+
+        self.bar_width = ceil(5 * scale_rater())
+
+        self.min_highlight = False  # min滑块是否高亮
+        self.max_highlight = False  # max滑块是否高亮
+
+        self.bind_min_handle = False  # 鼠标是否绑定min滑块
+        self.bind_max_handle = False  # 鼠标是否绑定max滑块
+
+        self.bar_color = Style().colors.light  # 拖动范围条的颜色
+        self.range_color = Color(Style().colors.light).set_brightness(1.5).hex  # 范围条的颜色
+        self.min_handle_base_color = Style().colors.primary  # 小值滑块基础颜色
+        self.max_handle_base_color = Color(Style().colors.primary).reverse().hex  # 大值滑块基础颜色
+        self.min_handle_color = copy(self.min_handle_base_color)  # 小值滑块颜色
+        self.max_handle_color = copy(self.max_handle_base_color)  # 大值滑块颜色
+
+        self.color_change()
+
+    def set(self, _min: float, _max: float):
+        self.min_percentage = _min
+        self.max_percentage = _max
+        self.redraw()
+
+    def color_change(self, *_):
+        colors = Style().colors
+        self.min_handle_base_color = colors.primary
+        self.max_handle_base_color = Color(colors.primary).reverse().hex
+        if Color(colors.bg).sum / 3 > 0.6:
+            self.bar_color = colors.light
+            self.range_color = Color(self.bar_color).set_brightness(1 / 1.5).hex
+        else:
+            self.bar_color = Color(colors.selectbg).set_brightness(0.8).hex
+            self.range_color = Color(self.bar_color).set_brightness(1.5).hex
+
+        self.update_color()
+        self.redraw()
+
+    def redraw(self, *_):
+        height = self.winfo_height()
+        width = self.winfo_width()
+
+        source = Image.new("RGBA", (self.winfo_width(), height), Style().colors.bg)
+        source2 = Image.new("RGBA", (self.winfo_width() * 10, height * 10), "#FFFF0000")
+
+        draw = ImageDraw.Draw(source)
+        draw2 = ImageDraw.Draw(source2)
+        draw.line((0, height // 2, self.winfo_width(), height // 2),
+                  fill=self.bar_color,
+                  width=self.bar_width)  # 绘制范围基条
+        draw.line((int(self.min_offset), height // 2, int(self.max_offset), height // 2),
+                  fill=self.range_color,
+                  width=self.bar_width)  # 绘制范围条
+        draw2.ellipse((int(self.min_offset) * 10, 0, (int(self.min_offset) + height) * 10, height * 10),
+                      fill=self.min_handle_color,
+                      width=0)  # 绘制小端滑块
+        draw2.ellipse((int(self.max_offset) * 10, 0, (int(self.max_offset) + height) * 10, height * 10),
+                      fill=self.max_handle_color,
+                      width=0)  # 绘制大端滑块
+
+        source2 = source2.resize((width, height))
+        source3 = Image.alpha_composite(source, source2)
+        self.image = ImageTk.PhotoImage(source3)
+        self.delete(ALL)
+        self.create_image(0, 0, anchor=NW, image=self.image)
+
+    def mouse_move(self, event: tk.Event):
+        min_box = (self.min_offset, 0, self.min_offset + 15, 15)
+        value = min_box[0] < event.x < min_box[2] and min_box[1] < event.y < min_box[3]
+        if value != self.min_highlight:
+            self.min_highlight = value
+            self.update_color()
+
+        max_box = (self.max_offset, 0, self.max_offset + 15, 15)
+        value = max_box[0] < event.x < max_box[2] and max_box[1] < event.y < max_box[3]
+        if value != self.max_highlight:
+            self.max_highlight = value
+            self.update_color()
+
+        if self.bind_min_handle:  # min滑块移动逻辑
+            self.min_percentage = (event.x - 7.5) / (self.winfo_width() - 15)
+            if self.min_percentage < 0:
+                self.min_percentage = 0
+            if self.min_percentage > self.max_percentage:
+                self.max_percentage = self.min_percentage
+                if self.max_percentage > 1:
+                    self.max_percentage = self.min_percentage = 1
+            self.event_generate("<<RangeChanged>>")
+            self.redraw()
+        elif self.bind_max_handle:  # max滑块移动逻辑
+            self.max_percentage = (event.x - 7.5) / (self.winfo_width() - 15)
+            if self.max_percentage > 1:
+                self.max_percentage = 1
+            if self.max_percentage < self.min_percentage:
+                self.min_percentage = self.max_percentage
+                if self.min_percentage < 0:
+                    self.min_percentage = self.max_percentage = 0
+            self.event_generate("<<RangeChanged>>")
+            self.redraw()
+
+    def update_color(self):
+        if self.min_highlight:
+            self.min_handle_color = Color(self.min_handle_base_color).set_brightness(1.1).hex
+        else:
+            self.min_handle_color = copy(self.min_handle_base_color)
+
+        if self.max_highlight:
+            self.max_handle_color = Color(self.max_handle_base_color).set_brightness(1.1).hex
+        else:
+            self.max_handle_color = copy(self.max_handle_base_color)
+        self.redraw()
+
+    def mouse_down(self, *_):
+        if self.max_highlight:
+            self.bind_max_handle = True
+        elif self.min_highlight:
+            self.bind_min_handle = True
+
+    def mouse_up(self, *_):
+        if self.bind_min_handle:
+            self.bind_min_handle = False
+        elif self.bind_max_handle:
+            self.bind_max_handle = False
+
+    @property
+    def value(self) -> (float, float):
+        return self.min_percentage, self.max_percentage
+
+    @property
+    def min_offset(self) -> float:
+        return (self.winfo_width() - self.winfo_height()) * self.min_percentage
+
+    @property
+    def max_offset(self) -> float:
+        return (self.winfo_width() - self.winfo_height()) * self.max_percentage
+
+
+class TipEntry(ttk.Entry):
+    def __init__(self, master: Misc, tip: str = "在此输入内容"):
+        super(TipEntry, self).__init__(master)
+        self.on_tip = False
+        self.tip = tip
+
+        self.bind("<FocusIn>", self.on_focus_get)
+        self.bind("<FocusOut>", self.on_focus_out)
+
+    def on_focus_get(self, *_):
+        if self.on_tip:
+            self.delete(0, END)
+            self.on_tip = False
+
+    def on_focus_out(self, *_):
+        if not self.get():
+            self.insert(0, self.tip)
+            self.on_tip = True
+
+    def set_tip(self, tip: str):
+        self.tip = tip
+
+
+class ProgressBar(ttk.Canvas):
+    def __init__(self, master: Misc, text: str = "0%"):
+        super(ProgressBar, self).__init__(master, height=26)
+        self.color = Style().colors
+        self.bind("<Configure>", self.redraw)
+        self.bind("<<ThemeChanged>>", self.change_color)
+        self.percentage = 0
+        self.text = text
+        self.redraw_lock = Lock()
+        self.now_elements = []
+        self.last_elements = []
+
+        self.text_id = 0
+        self.redraw()
+
+    def change_color(self, *_):
+        self.color = Style().colors
+        self.redraw()
+
+    def redraw(self, *_):
+        with self.redraw_lock:
+            self.now_elements.clear()
+            width = self.winfo_width()
+            bar_x = int((width - 4) * self.percentage)
+            if bar_x == 1:
+                self.now_elements.append(self.create_line(1, 1, 1, 26, fill=self.color.success))
+            elif bar_x > 1:
+                self.now_elements.append(self.create_rectangle(1, 1, bar_x, 26 - 2, fill=self.color.success,
+                                                               outline=self.color.success))
+
+            self.now_elements.append(self.create_rectangle(0, 0, width - 1, 26 - 1, outline=self.color.border))
+            self.text_id = self.create_text(width // 2, 13, text=self.text, fill=self.color.fg)
+            self.now_elements.append(self.text_id)
+
+            if self.last_elements:
+                self.delete(*self.last_elements)
+            self.last_elements = self.now_elements.copy()
+
+    def set_percentage(self, percentage: float, text: str = None):
+        self.percentage = percentage
+        self.text = text if text is not None else self.text
+        self.redraw()
+
+
 def get_now_time() -> str:
     return strftime("%Y-%m-%d_%H-%M-%S", localtime())
-
-
-def set_default_font():
-    font.nametofont("TkDefaultFont").config(family="微软雅黑", size=10)
-
-
-def load_unifont():
-    from pyglet import options
-    from pyglet.font import add_file
-
-    if not exists("assets/Unifont.otf"):  # 若字体文件不存在则退出
-        print("Unifont字体文件丢失", file=stderr)
-        return
-
-    options['win32_gdi_font'] = True
-    add_file("assets/Unifont.otf")
 
 
 def write_msg_window_buttons(left: str, right: str, timeout: float = 1.2):
@@ -83,55 +435,6 @@ def write_msg_window_buttons(left: str, right: str, timeout: float = 1.2):
             return
         EnumChildWindows(msg_win, callback, None)
         break
-
-
-class GUI(ttk.Window):
-    def __init__(self):
-        global scanbar
-        super(GUI, self).__init__()
-
-        timer = perf_counter()
-
-        set_default_font()
-        self.config_root_window()
-
-        self.title_bar = TitleBar(self)
-        self.sep = ttk.Separator()
-        self.tabs = Tabs(self)
-        self.logger = Logger(self.tabs)
-        self.server_scanF = ttk.Frame(self.tabs)
-        self.servers = ServerList(self.server_scanF, self.logger)
-        self.scan_bar = scanbar = ScanBar(self.server_scanF, self.logger, self.servers)
-
-        self.pack_widgets()
-        print(f"GUI构建时间: {perf_counter() - timer:.3f}秒")
-        Thread(target=load_unifont).start()  # 加载字体
-
-    def config_root_window(self):  # 设置窗体
-        self.wm_title("MC服务器扫描器")  # 设置标题
-        self.style.theme_use("solar")
-        self.protocol("WM_DELETE_WINDOW", self.on_delete_window)
-        Thread(target=self.set_icon).start()
-        Thread(target=self.place_window_center).start()
-
-    def set_icon(self):
-        if exists("assets/icon.ico"):
-            self.wm_iconbitmap("assets/icon.ico")
-        else:
-            print("图标文件丢失", file=stderr)
-
-    def on_delete_window(self):
-        self.scan_bar.close_save_config()
-        self.destroy()
-
-    def pack_widgets(self):
-        self.title_bar.pack(fill=X, padx=10)
-        self.sep.pack(fill=X, padx=10, pady=3)
-        self.tabs.pack(fill=BOTH, expand=True, pady=0)
-        self.servers.pack(fill=BOTH, expand=True, padx=3, pady=3)
-        self.scan_bar.pack(side=BOTTOM, fill=X, padx=3, pady=3)
-        self.tabs.add(self.server_scanF, text="控制面板")
-        self.tabs.add(self.logger, text="日志")
 
 
 class ServersFilter:
@@ -556,7 +859,7 @@ class ServerFrame(ttk.Frame):
         if self.data.has_favicon:
             self.favicon.configure(image=self.data.favicon_photo)
         else:
-            self.default_favicon = tk.PhotoImage(master=self, file=r"assets/server_icon.png")
+            self.default_favicon = tk.PhotoImage(master=self, file=r"../assets/server_icon.png")
             self.favicon.configure(image=self.default_favicon)
 
         self.MOTD.load_motd(self.data)
@@ -580,7 +883,7 @@ class ServerFrame(ttk.Frame):
         self.info_window: InfoWindow = InfoWindow(self, self.data)
 
     def load_motd_text(self):
-        text = ""
+        text = str()
         for extra in self.data.description_json:
             text += extra["text"]
         return text
@@ -657,54 +960,6 @@ class InfoProgressBar(ttk.Frame):
         self.speed_avg.append((value - self.last_value) / (time() - self.last_update))
         self.last_value = value
         self.last_update = time()
-
-    def finish(self):
-        self.update_now(self.max_)
-        self.update_progress_text(0)
-
-
-class ProgressBar(ttk.Canvas):
-    def __init__(self, master: Misc, text: str = "0%"):
-        super(ProgressBar, self).__init__(master, height=26)
-        self.color = Style().colors
-        self.bind("<Configure>", self.redraw)
-        self.bind("<<ThemeChanged>>", self.change_color)
-        self.percentage = 0
-        self.text = text
-        self.redraw_lock = Lock()
-        self.now_elements = []
-        self.last_elements = []
-
-        self.text_id = 0
-        self.redraw()
-
-    def change_color(self, *_):
-        self.color = Style().colors
-        self.redraw()
-
-    def redraw(self, *_):
-        with self.redraw_lock:
-            self.now_elements.clear()
-            width = self.winfo_width()
-            bar_x = int((width - 4) * self.percentage)
-            if bar_x == 1:
-                self.now_elements.append(self.create_line(1, 1, 1, 26, fill=self.color.success))
-            elif bar_x > 1:
-                self.now_elements.append(self.create_rectangle(1, 1, bar_x, 26 - 2, fill=self.color.success,
-                                                               outline=self.color.success))
-
-            self.now_elements.append(self.create_rectangle(0, 0, width - 1, 26 - 1, outline=self.color.border))
-            self.text_id = self.create_text(width // 2, 13, text=self.text, fill=self.color.fg)
-            self.now_elements.append(self.text_id)
-
-            if self.last_elements:
-                self.delete(*self.last_elements)
-            self.last_elements = self.now_elements.copy()
-
-    def set_percentage(self, percentage: float, text: str = None):
-        self.percentage = percentage
-        self.text = text if text is not None else self.text
-        self.redraw()
 
 
 class PauseButton(ttk.Button):
@@ -895,7 +1150,7 @@ class ScanBar(ttk.LabelFrame):
         self.callback_workers = 0
         self.taskbar = None
         self.config_path = path_join(config_dir, "scan_config.json")
-        self.user_address_operator = vars.UserAddressOperator()
+        self.user_address_operator = UserAddressOperator()
         Thread(target=self.taskbar_create, daemon=True).start()
 
         # 进度条
@@ -908,7 +1163,7 @@ class ScanBar(ttk.LabelFrame):
 
         # 输入 Frame
         self.input_frame = ttk.Frame(self)
-        self.host_input = TextCombobox(self.input_frame, "域名: ", vars.server_addresses)
+        self.host_input = TextCombobox(self.input_frame, "域名: ", server_addresses)
         self.timeout_input = EntryScaleFloat(self.input_frame, 0.1, 3.0, 0.2, "超时时间: ")
         self.thread_num_input = EntryScaleInt(self.input_frame, 1, 256, 192, "线程数: ")
         self.range_input = RangeSelector(self.input_frame, "端口选择: ", 1024, 65535)
@@ -970,17 +1225,10 @@ class ScanBar(ttk.LabelFrame):
         timer = perf_counter()
         main_window = 0  # 默认值为0会使TaskbarApi可调用但无效
         while perf_counter() - timer < timeout:
-            hwnd = self.winfo_id()
-            while True:
-                parent = GetParent(hwnd)
-                if parent == 0:
-                    break
-                else:
-                    hwnd = copy(parent)
-            if GetWindowText(hwnd) == "MC服务器扫描器" and GetClassName(hwnd) == "TkTopLevel":
-                main_window = copy(hwnd)
+            root_hwnd = GetParent(self.master.master.master.winfo_id())
+            if GetWindowText(root_hwnd) == "MC服务器扫描器" and GetClassName(root_hwnd) == "TkTopLevel":
+                main_window = root_hwnd
                 break
-            sleep(0.05)
         else:
             print("main_gui.taskbar_create: Main Window Not Found!", file=stderr)
         if main_window != 0:
@@ -993,7 +1241,7 @@ class ScanBar(ttk.LabelFrame):
         with self.callback_lock:
             self.progress_var += 1
             self.progress_bar.update_progress(self.progress_var)
-            self.taskbar.SetProgressValue(self.progress_var, self.progress_bar.max_)
+            self.taskbar.set_progress_value(int(self.progress_var / self.progress_bar.max_ * 100), 100)
             if isinstance(info, ServerInfo):
                 self.server_list.add_server(info)
                 self.logger.log(INFO, f"[{info.port}]:", "检测到MC服务器")
@@ -1024,12 +1272,12 @@ class ScanBar(ttk.LabelFrame):
         self.scan_obj.config(timeout, thread_num, self.callback)
         Thread(target=self.scan_obj.run, args=(host, range(start, stop))).start()
         Thread(target=self.check_over_thread, daemon=True).start()
-        self.taskbar.SetProgressState(TBPFLAG.TBPF_NORMAL)
+        self.taskbar.set_progress_state(TBPFLAG.TBPF_NORMAL)
 
         # 写入配置文件，使得下一次自动加载
         self.logger.log(INFO, f"将地址 [{host}] 写入配置文件。")
-        writeResult = self.user_address_operator.writeAddressToConfigFile(address=host)
-        if writeResult is False:
+        write_result = self.user_address_operator.writeAddressToConfigFile(address=host)
+        if write_result is False:
             self.logger.log(ERROR, f"写入地址 [{host}] 时，文件操作时发生错误！")
             MessageBox(self.winfo_id(),
                        f"对：{self.user_address_operator.user_address_json} 文件操作时发生错误！",
@@ -1048,7 +1296,7 @@ class ScanBar(ttk.LabelFrame):
             self.pause_button.configure(state=NORMAL)
 
         Thread(target=task, daemon=True).start()
-        self.taskbar.SetProgressState(TBPFLAG.TBPF_PAUSED)
+        self.taskbar.set_progress_state(TBPFLAG.TBPF_PAUSED)
 
     def resume_scan(self):
         def task():
@@ -1061,7 +1309,7 @@ class ScanBar(ttk.LabelFrame):
             self.pause_button.configure(state=NORMAL)
 
         Thread(target=task, daemon=True).start()
-        self.taskbar.SetProgressState(TBPFLAG.TBPF_NORMAL)
+        self.taskbar.set_progress_state(TBPFLAG.TBPF_NORMAL)
 
     def stop_scan(self):
         if not self.in_scan:
@@ -1073,18 +1321,18 @@ class ScanBar(ttk.LabelFrame):
         def stop_task():
             self.in_scan = False
             self.logger.log(DEBUG, "停止扫描")
-            self.taskbar.SetProgressState(TBPFLAG.TBPF_ERROR)
+            self.taskbar.set_progress_state(TBPFLAG.TBPF_ERROR)
             self.scan_obj.stop()
             while self.scan_obj.worker_count > 0 or self.scan_obj.callback_count > 0:
                 sleep(0.1)
                 self.logger.log(DEBUG, "等待工作线程全部结束, 剩余数量:", self.scan_obj.worker_count)
             self.logger.log(DEBUG, "工作线程已全部结束")
             self.start_button.configure(state=NORMAL)
-            self.taskbar.SetProgressState(TBPFLAG.TBPF_NOPROGRESS)
+            self.taskbar.set_progress_state(TBPFLAG.TBPF_NOPROGRESS)
             self.progress_stop()
 
         Thread(target=stop_task).start()
-        self.taskbar.SetProgressState(TBPFLAG.TBPF_INDETERMINATE)
+        self.taskbar.set_progress_state(TBPFLAG.TBPF_INDETERMINATE)
 
     def check_over_thread(self):
         self.logger.log(DEBUG, "检测扫描结束线程启动...")
@@ -1105,8 +1353,8 @@ class ScanBar(ttk.LabelFrame):
             self.start_button.configure(state=NORMAL)
             self.progress_stop()
             self.progress_bar.update_progress(self.progress_bar.max_)  # 掩耳盗铃一波
-            FlashWindowCount(self.taskbar.hwnd, 3, 0.3)
-            self.taskbar.SetProgressState(TBPFLAG.TBPF_NOPROGRESS)
+            self.taskbar.flash_window()
+            self.taskbar.set_progress_state(TBPFLAG.TBPF_NOPROGRESS)
 
     def progress_stop(self):
         self.progress_bar.update_now(self.progress_var)
